Moved the language definition for moment.js to src/config and updated the documentation
