+++ src/config/lang-phrases/moment-am.js
